web/requirements.txt
Flask
flask_restful
pymongo
bcrypt
python-dotenv
pymupdf  # install PyMuPDF but import as fitz
nltk
openai
